Move corgit functionality into server; add issue titles to PR description
