+++ server.py
@@ -11,14 +11,18 @@ import tornado.template
 
 from jenkinsapi.jenkins import Jenkins
 
+import github
+
 from corgi import Corgi
-from corgit import update_pr_description, get_issues_from_pr, get_pullrequest
 from config import config
 
 from logging import StreamHandler
 from logging.handlers import WatchedFileHandler
 
 log = logging.getLogger('server')
+
+
+HEADER = '### Referenced Issues:'
 
 
 def create_tree_url(data):
@@ -73,6 +77,64 @@ def run_jenkins_job(job):
     else:
         logging.error('Jenkins job %s not found' % job)
         logging.debug('Available Jenkins jobs: %s' % ', ' % jenkins.keys())
+
+
+def get_pullrequest(repo_name, pr_number):
+    gh = github.Github(config['git.token'])
+    repo = gh.get_repo(repo_name)
+    return repo.get_pull(pr_number)
+
+
+def get_issues_from_pr(pullrequest):
+    text = [pullrequest.title, pullrequest.body]
+    for commit in pullrequest.get_commits():
+        text.append(commit.commit.message)
+    return sorted(set(map(int, re.findall(r'\bgs-(\d+)', ' '.join(text)))))
+
+
+def get_issue_titles(issues):
+    corgi = Corgi(config['redmine.url'], config['redmine.auth_key'])
+    titles = dict()
+    if corgi.connected:
+        for issue in issues:
+            titles[issue] = corgi.getIssueTitle(issue)
+    return titles
+
+
+def update_pr_description(pullrequest, dryrun=False):
+    log.info('Updating PR description for %s PR %s' % (pullrequest.base.repo.full_name, pullrequest.number))
+    body = pullrequest.body
+    issues = get_issues_from_pr(pullrequest)
+    titles = get_issue_titles(issues)
+    links = '\n'.join('* [Issue %s: %s](%sissues/%s)' % (issue, titles[issue], config['redmine.url'], issue)
+                      for issue in issues)
+    lines = [line.strip() for line in body.split('\n')]
+    if HEADER in lines:
+        log.info('Found existing list of issues, updating')
+        # update existing list
+        pos = lines.index(HEADER) + 1
+        while pos < len(lines) and lines[pos].startswith('* '):
+            del lines[pos]
+        if links:
+            lines.insert(pos, links)
+        else:
+            log.info('Removing existing list of issues')
+            del lines[pos - 1]
+    elif links:
+        log.info('No existing list of issues found, creating')
+        lines.append(HEADER)
+        lines.append(links)
+
+    updated_body = '\n'.join(lines)
+
+    if updated_body != body:
+        log.info('Committing new body')
+        if not dryrun:
+            pullrequest.edit(body=updated_body)
+    else:
+        log.info('Body unchanged, skipping commit')
+
+    return updated_body
 
 
 class EventHandler(tornado.web.RequestHandler):
